+++ README.md
@@ -1,14 +1,14 @@
-# Text-to-Image Diffusion Benchmark
+# Holistic Evaluation of Multi-modal Generative Models
 
-`t2i_diffusion_benchmark` is a library for performing comprehensive benchmark of text-to-image diffusion models on image quality and prompt comprehension integrated with [Weights & Biases](https://wandb.ai/site) and [Weave](https://wandb.github.io/weave/).
+`hemm` is a library for performing comprehensive benchmark of text-to-image diffusion models on image quality and prompt comprehension integrated with [Weights & Biases](https://wandb.ai/site) and [Weave](https://wandb.github.io/weave/). `hemm` is inspired by [Holistic Evaluation of Text-To-Image Models](https://crfm.stanford.edu/helm/heim/v1.0.0/).
 
 ```python
-from t2i_diffusion_benchmark.metrics import (
+from hemm.metrics import (
     CLIPImageQualityScorer,
     CLIPScorer,
     BLIPScorer,
 )
-from t2i_diffusion_benchmark.eval_pipelines import StableDiffusionEvaluationPipeline
+from hemm.eval_pipelines import StableDiffusionEvaluationPipeline
 
 
 diffuion_evaluation_pipeline = StableDiffusionEvaluationPipeline(
