+++ hemm/eval_pipelines/stable_diffusion_evaluation.py
