+++ hemm/metrics/prompt_alignment/blip_scorer.py
