+++ hemm/metrics/prompt_alignment/clip_iqa_scorer.py
