+++ hemm/metrics/prompt_alignment/clip_scorer.py
