+++ pyproject.toml
@@ -1,10 +1,10 @@
 [tool.poetry]
-name = "t2i_diffusion_benchmark"
+name = "hemm"
 version = "0.1.0"
 description = ""
 authors = ["soumik12345 <19soumik.rakshit96@gmail.com>"]
 readme = "README.md"
-packages = [{ include = "t2i_diffusion_benchmark" }]
+packages = [{ include = "hemm" }]
 
 [tool.poetry.dependencies]
 python = "^3.10"
